pages/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
huggingface_hub==0.16.4
sentence-transformers==2.2.2
numpy==1.26.2
pydantic==2.5.2
python-multipart==0.0.6 